% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interaction.preferences}
\alias{interaction.preferences}
\title{Estimate species' interaction preferences based on an interaction matrix and their relative abundances}
\usage{
interaction.preferences(adj, row.abun = NULL, col.abun = NULL,
  include.zeros = FALSE)
}
\arguments{
\item{adj}{The adjacency matrix}

\item{row.abun}{The abundances (or relative abundances) of row species}

\item{col.abun}{The abundances (or relative abundances) of column species}

\item{include.zeros}{Flag to decide whether or not unobserved interactions should have their preferences estimated}
}
\description{
Estimate species' interaction preferences based on an interaction matrix and their relative abundances
}

